\name{sibships-package}
\alias{sibships-package}
\alias{sibships}
\docType{package}
\title{
\packageTitle{sibships}
}
\description{
\packageDescription{sibships}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{sibships}
\packageIndices{sibships}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{sibships}

Maintainer: \packageMaintainer{sibships}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
